requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
pillow==10.2.0
faker
gunicorn
whitenoise
dj-database-url
python-dotenv
